Flask/requirements.txt
Flask
translate
easyocr
Pillow
